Main Page RecyclerView and Calendar Model Usage
Made the main page functional (along with its RecyclerView), as well as having the models be put to use.

Major fixes to the models
Minor fixes to Log In error checking
Main page is now dynamic, in a sense that it displays the current month and year of the local device

Next goal is to effectively switch displayed month (and year, kung lalagpas ba tayong 2021).