Notifications (finally working)
Add an event as usual. To set an alarm for an event, go back to the main event page, click on "SET ALARMS", find your desired event and click on it. A Snackbar will pop-up to notify you of the alarm set, of what time and of what event. Now works with multiple alarms.